+++ EasySaver/src/AppName.cs
@@ -1,0 +1,10 @@
+﻿namespace EasySaver.Common
+{
+    public partial class EasySaver
+    {
+        /// <summary>
+        /// Application's name.
+        /// </summary>
+        public static readonly string AppName = System.Reflection.Assembly.GetExecutingAssembly().GetName().Name;
+    }
+}