+++ EasySaver/src/DefaultExtension.cs
@@ -1,0 +1,11 @@
+﻿namespace EasySaver.Common
+{
+    public partial class EasySaver
+    {
+        // Default text file extension.
+        internal static string s_defaultTextExtension = ".txt";
+
+        // Default image file extension.
+        internal static string s_defaultImageExtension = ".bmp";
+    }
+}