+++ EasySaver/src/Path.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace EasySaver.Common
+{
+    public partial class EasySaver
+    {
+        /// <summary>
+        /// Desktop path.
+        /// </summary>
+        public static readonly string DesktopPath = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
+    }
+}